obama

+++ src/comands/dj.ts
@@ -31,8 +31,14 @@ export async function dj(bot: Ibot, nome: string) {
         }
     }
     console.log(`complete`)
-    const image = await downloadAxios(result[0].thumbnail, "png")
+    if(result[0]){
+        const image = await downloadAxios(result[0].thumbnail, "png")
+        await sendImage(image, `🎧 melhores musicas de _${nome}_ 🎧`, true)
+        return fs.unlinkSync(image)
+    }else{
+        await reply(`🎧 melhores musicas de _${nome}_ 🎧`)
+
+    }
+    
   
-    await sendImage(image, `🎧 melhores musicas de _${nome}_ 🎧`, true)
-    return fs.unlinkSync(image)
 }
